:truck: Give data sheets to appropriate names
These data sheets and user manuals are not the only ones for the stm32f4
family and as such should be renamed to reflect that. It may even be
appropriate to determine if the stm32f411 and stm32f40x series are
compatible enough to warrant them being in the same package OR if they
should get separate packages.